backend/pyproject.toml
[project]
name = "recipe-companion-backend"
version = "0.1.0"
description = "Recipe analysis agent with CopilotKit integration"
requires-python = ">=3.11"
dependencies = [
    "copilotkit>=0.1.74",
    "pydantic-ai>=0.1.0",
    "ag-ui-protocol>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.18",
    "pypdf>=5.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",

]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ipdb>=0.13.13",
    "ruff>=0.14.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests that require real API calls (deselect with '-m \"not integration\"')",
]
